@@ -1,69 +1,63 @@
 {
-  "name": "FingerNFT",
+  "name": "finger-nft-web",
   "version": "1.0.1",
-  "description": "Finger NFT",
-  "main": "src/main.js",
   "author": "",
   "contributors": [],
   "private": false,
   "scripts": {
-    "dev": "vue-cli-service serve",
-    "build": "node --max_old_space_size=4096 ./node_modules/@vue/cli-service/bin/vue-cli-service.js build --mode production",
+    "dev": "cross-env NODE_ENV=development vite",
+    "build": "node --max_old_space_size=4096 ./node_modules/@vue/cli-service/bin/vue-cli-service.js build --mode prproductionod",
     "serve": "vue-cli-service serve",
     "start": "vue-cli-service serve --open"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ],
   "dependencies": {
-    "@bugsnag/js": "^7.10.5",
-    "@bugsnag/plugin-vue": "^7.10.0",
+    "@bugsnag/js": "^8.0.0",
+    "@bugsnag/plugin-vue": "^8.0.0",
     "@soerenmartius/vue3-clipboard": "^0.1.2",
-    "axios": "^0.21.1",
-    "bignumber.js": "7.2.1",
-    "dotenv": "^8.0.0",
-    "ethereumjs-util": "^7.0.10",
+    "axios": "^1.7.7",
+    "bignumber.js": "^9.1.2",
+    "dotenv": "^16.4.5",
+    "ethereumjs-util": "^7.1.5",
     "ethers": "^5.1.0",
     "lib-flexible": "^0.3.2",
     "mersenne-twister": "^1.1.0",
     "normalize.css": "^8.0.1",
-    "vue": "^3.2.0",
-    "vue-router": "^4.0.0-0",
-    "vuex": "^4.0.0-0",
-    "vuex-persist": "3.1.3"
+    "vue": "^3.4.38",
+    "vue-router": "^4.4.5",
+    "vuex": "^4.1.0",
+    "vuex-persist": "^3.1.3",
+    "buffer": "^6.0.3"
   },
   "devDependencies": {
-    "@truffle/contract": "^4.3.14",
-    "@vue/cli-plugin-babel": "4.5.0",
-    "@vue/cli-plugin-pwa": "4.5.0",
-    "@vue/cli-plugin-router": "4.5.0",
-    "@vue/cli-plugin-vuex": "4.5.0",
-    "@vue/cli-service": "^4.5.0",
-    "@vue/compiler-sfc": "3.0.0",
-    "bn.js": "^5.2.0",
-    "compression-webpack-plugin": "^6.1.1",
-    "core-js": "^3.10.2",
-    "element-plus": "^1.1.0-beta.24",
-    "element-theme": "^2.0.1",
+    "@vitejs/plugin-vue": "^5.1.3",
+    "@truffle/contract": "^4.6.31",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-pwa": "^5.0.8",
+    "@vue/cli-plugin-router": "^5.0.8",
+    "@vue/cli-plugin-vuex": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/compiler-sfc": "^3.5.5",
+    "@rollup/plugin-inject": "^5.0.5",
+    "bn.js": "^5.2.1",
+    "core-js": "^3.38.1",
+    "element-plus": "^2.8.3",
     "element-theme-chalk": "^2.15.1",
+    "color": "^4.2.3",
     "ethereumjs-abi": "^0.6.8",
-    "lint-staged": "9.5.0",
-    "node-sass": "^5.0.0",
-    "prettier": "1.19.1",
-    "sass": "^1.32.11",
-    "sass-loader": "^10.1.1",
-    "typescript": "3.9.3",
-    "vue-i18n": "^9.1.6",
-    "vue-loader-v16": "^16.0.0-beta.5.4",
-    "web3": "^1.5.0",
-    "webpack": "^4.44.2"
-  },
-  "engines": {
-    "node": ">= 8.0.0",
-    "npm": ">= 5.0.0",
-    "yarn": ">= 1.0.0"
-  },
-  "keywords": []
+    "lint-staged": "^15.2.10",
+    "prettier": "^3.3.3",
+    "sass": "^1.78.0",
+    "sass-loader": "^16.0.1",
+    "typescript": "^5.6.2",
+    "vue-i18n": "^10.0.1",
+    "web3": "^4.12.1",
+    "vite": "^5.4.5",
+    "vite-plugin-components": "^0.11.5",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-mock": "^3.0.0",
+    "vite-plugin-pages": "^0.31.0",
+    "cross-env": "^7.0.3",
+    "mockjs": "^1.1.0",
+    "vite-plugin-node-polyfills": "^0.22.0"
+  }
 }
