@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   global: {
     notConnect: "钱包未连接",
     connect: "连接钱包?",
